+++ !Scrabble/game/scrabble.py
@@ -2,3 +2,12 @@ from game.tile import Tile
 from game.bag_tiles import BagTiles
 from game.player import Player
 from game.square import Square
+from game.board import Board
+
+class ScrabbleGame:
+    def __init__(self, players_count):
+        self.board = Board()
+        self.bag_tiles = BagTiles()
+        self.players = []
+        for _ in range(players_count):
+            self.players.append(Player())