+++ !Scrabble/game/square.py
@@ -18,4 +18,3 @@ class Square:
             return self.letter.value * self.multiplier
         else:
             return self.letter.value
-